added grpc service functionality

+++ internal/base/base.go
@@ -6,19 +6,19 @@ func NewSimpleMathService() *SimpleMathService {
 	return &SimpleMathService{}
 }
 
-func (s *SimpleMathService) Add(a, b int32) int32 {
+func (s *SimpleMathService) Add(a, b float32) float32 {
 	return a + b
 }
 
-func (s *SimpleMathService) Sub(a, b int32) int32 {
+func (s *SimpleMathService) Sub(a, b float32) float32 {
 	return a - b
 }
 
-func (s *SimpleMathService) Mul(a, b int32) int32 {
+func (s *SimpleMathService) Mul(a, b float32) float32 {
 	return a * b
 }
 
-func (s *SimpleMathService) Div(a, b int32) int32 {
+func (s *SimpleMathService) Div(a, b float32) float32 {
 	if b == 0 {
 		return 0
 	}
