+++ internal/core/types/types.go
@@ -5,8 +5,8 @@ type Server interface {
 }
 
 type MathService interface {
-	Add(a, b int32) int32
-	Sub(a, b int32) int32
-	Mul(a, b int32) int32
-	Div(a, b int32) int32
+	Add(a, b float32) float32
+	Sub(a, b float32) float32
+	Mul(a, b float32) float32
+	Div(a, b float32) float32
 }
